requirements.txt
streamlit
pandas
python-dotenv
openai>=1.30.0
